Switch to generic oauth setup for extension that require auth.
The extension registry now supports the following metadata:

oauth:
  button: URL for the image to use for the button.
  target: URL to point the popup at for the OAuth flow.
  config:
    section_name
      - field_name
  note: Extra markdown text to add at the end of the section.

JS code in mopidy.js will wire up a click handler for the button which launches
the popup. Once this is up we monitor the popups state and wait for the callback
to post a message with the result.

Only messages from https://mopidy.com and https://auth.mopidy.com are supported.
The message is expected to be just key values that match the configured field
names. E.g. `auth_token` for Soundcloud. `error` and `error_description` should
be set when the authentication failed.

It's the callbacks job to post the message back to `window.opener` and make sure
the field names match.