Create autoplay.md
Hi, I wrote (and uploaded this new extension to PyPi), because I hated it from the beginning, that after switching on my Raspberry running Mopidy I had to start the computer, start the browser and head to Mopidy to start playing my favourite webradio.

I would appreciate, if you could add it to your list of extensions (or give me some comments about the implementation ;-) ). Thanks.